requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.115.3
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
pydantic==2.9.2
pydantic_core==2.23.4
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.2
selenium==4.25.0
slack_bolt==1.21.1
slack_sdk==3.33.1
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.41.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
websocket-client==1.8.0
wsproto==1.2.0
